requirements.txt
python-telegram-bot
Flask
tinydb